+++ src/MultiFactor.Radius.Adapter.Tests/Data/UsernameTransformationRules/UsernameTransformationRuleTestCase.cs
@@ -3,11 +3,11 @@ namespace MultiFactor.Radius.Adapter.Tests.Data.UsernameTransformationRules
 {
     public class UsernameTransformationRuleTestCase
     {
-        public string Asset = "";
-        public string ReplaceFirst = "";
-        public string MatchFirst = "";
-        public string ReplaceSecond = "";
-        public string MatchSecond = "";
+        public string Asset { get; init; } = "";
+        public string ReplaceFirst { get; init; } = "";
+        public string MatchFirst { get; init; } = "";
+        public string ReplaceSecond { get; init; } = "";
+        public string MatchSecond { get; init;  } = "";
     }
 
 }
