+++ src/MultiFactor.Radius.Adapter/Infrastructure/Configuration/ClientLevel/ClientConfiguration.cs
@@ -7,8 +7,7 @@ using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.Authentic
 using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.PreAuthModeFeature;
 using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.PrivacyModeFeature;
 using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.RandomWaiterFeature;
-using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.UserNameTransformFeature;
-using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Models.UserNameTransform;
+using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.UserNameTransform;
 using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Models.UserNameTransformFeature;
 
 namespace MultiFactor.Radius.Adapter.Infrastructure.Configuration.ClientLevel;
