+++ src/MultiFactor.Radius.Adapter/Infrastructure/Configuration/ClientLevel/IClientConfiguration.cs
@@ -5,7 +5,7 @@ using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.Authentic
 using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.PreAuthModeFeature;
 using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.PrivacyModeFeature;
 using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.RandomWaiterFeature;
-using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Models.UserNameTransform;
+using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.UserNameTransform;
 
 namespace MultiFactor.Radius.Adapter.Infrastructure.Configuration.ClientLevel;
