+++ src/MultiFactor.Radius.Adapter/Infrastructure/Configuration/Features/UserNameTransform/UserNameTransformRulesScope.cs
@@ -1,4 +1,4 @@
-﻿namespace MultiFactor.Radius.Adapter.Infrastructure.Configuration.Models.UserNameTransform;
+﻿namespace MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.UserNameTransform;
 
 public enum UserNameTransformRulesScope
 {
