+++ src/MultiFactor.Radius.Adapter/Infrastructure/Configuration/Models/UserNameTransform/UserNameTransformRulesSection.cs
@@ -2,7 +2,6 @@
 //Please see licence at 
 //https://github.com/MultifactorLab/multifactor-radius-adapter/blob/main/LICENSE.md
 using Microsoft.Extensions.Configuration;
-using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Features.UserNameTransformFeature;
 using MultiFactor.Radius.Adapter.Infrastructure.Configuration.Models.UserNameTransformFeature;
 
 namespace MultiFactor.Radius.Adapter.Infrastructure.Configuration.Models.UserNameTransform;
