+++ src/MultiFactor.Radius.Adapter/Program.cs
@@ -23,7 +23,7 @@ try
 
     builder.UseMiddleware<StatusServerMiddleware>();
     builder.UseMiddleware<AccessRequestFilterMiddleware>();
-    builder.UseMiddleware<TransformUserNameMiddleware>();
+    builder.UseMiddleware<FirstFactorTransformUserNameMiddleware>();
     builder.UseMiddleware<AccessChallengeMiddleware>();
     builder.UseMiddleware<AnonymousFirstFactorAuthenticationMiddleware>();
     builder.UseMiddleware<PreSecondFactorAuthenticationMiddleware>();
