+++ src/MultiFactor.Radius.Adapter/Server/Pipeline/TransformUserName/FirstFactorTransformUserNameMiddleware.cs
@@ -1,4 +1,4 @@
-﻿//Copyright(c) 2020 MultiFactor
+//Copyright(c) 2020 MultiFactor
 //Please see licence at 
 //https://github.com/MultifactorLab/multifactor-radius-adapter/blob/main/LICENSE.md
 
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace MultiFactor.Radius.Adapter.Server.Pipeline.TransformUserName
 {
-    public class SecondFactorTransformUserNameMiddleware : IRadiusMiddleware
+    public class FirstFactorTransformUserNameMiddleware : IRadiusMiddleware
     {
         public Task InvokeAsync(RadiusContext context, RadiusRequestDelegate next)
         {
